Working on level flow

+++ Monocle/MonocleEngine/Components/Logic/Tween.cs
@@ -154,7 +154,7 @@ namespace Monocle
         static public Tween Alpha(GraphicsComponent image, float targetAlpha, int duration, Ease.Easer easer, TweenMode tweenMode = TweenMode.Oneshot)
         {
             Entity entity = image.Entity;
-            float startAlpha = image.Color.A / 255;
+            float startAlpha = ((float)image.Color.A / 255);
             Tween tween = new Tween(tweenMode, easer, duration, true);
             tween.OnUpdate = (t) => { image.Color.A = (byte)Math.Round(MathHelper.Lerp(startAlpha, targetAlpha, t.Eased) * 255.0f); };
             entity.Add(tween);
